# api/openapi.yaml
openapi: 3.0.0
# https://support.smartbear.com/swaggerhub/docs/tutorials/openapi-3-tutorial.html
info:
  version: 1.0.0
  title: Mailbox API
  description: Mailbox App's API specification

servers:
  - url: https://0.0.0.0:3000

paths:
  /v1/users:
    post:
      tags: 
        - Users
      description: Creating a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#components/schemas/UserRequest'
      responses:
        '200':
          description: Successfully return created user's Id
          content:
            application/json:
              schema:
                type: string
        '500':
          $ref: '#components/responses/500Error'
  /v1/users/:id:
    get:
      tags:
        - Users
      description: Getting user by its Id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Return user info by Id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                type: string
        '500':
          $ref: '#components/responses/500Error'
  /v1/mailboxes:
    post:
      tags:
        - Mailboxes
      description: Creating a mailbox
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MailboxRequest'
      responses:
        '200':
          description: Successfully return created mailbox Id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MailboxResponse'
        '500':
          $ref: '#components/responses/500Error'
  /v1/mailboxes/user/:id:
    get:
      tags:
        - Mailboxes
      description: Getting list of mailboxes by user Id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Returns list of mailboxes by user's Id
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  $ref: '#/components/schemas/MailboxResponse'
        '404':
          description: Mailboxes not found
          content:
            application/json:
              schema:
                type: string
        '500':
          $ref: '#components/responses/500Error'
components:
  schemas:
    UserRequest:
      type: object
      properties:
        user_name:
          type: string
        password:
          type: string
      required:
        - "user_name"
        - "password"
    UserResponse:
      type: object
      properties:
        id:
          type: string
        user_name:
          type: string
    MailboxRequest:
      type: object
      properties:
        user_id:
          type: string
        email:
          type: string
    MailboxResponse:
      type: object
      properties:
        id:
          type: string
        user_id:
          type: string
        email:
          type: string
  responses:
    500Error:
      description: Return error in case of failure
      content:
        application/json:
          schema:
            type: string